+++ src/classes_hex.py
@@ -1,7 +1,8 @@
 import pygame
 import random
 
-from settings import *
+# from settings import *
+from setup import *
 from functions_math import *
 
 class HexTile: