+++ src/functions_player.py
@@ -1,4 +1,5 @@
-from settings import *
+# from settings import *
+from setup import *
 from functions_graphics import *
 from functions_math import *
 
@@ -55,8 +56,44 @@ def unit_selection(win, list_with_units, corner1, corner2, offset_x, offset_y, s
     pygame.draw.rect(win, LIME, rect, 3)
 
 
-def set_new_target(list_with_units, target, is_ctrl_down):
+def set_new_target_move(list_with_units, target, is_ctrl_down):
 # set new movement target to all selected units
+# units will move in original arrangement
+    number_of_selested_units = 0
+    # biggest_unit_radius = 0
+    slowest_unit_speed = 100
+    top_left = 99999
+    top_right = -99999
+    top_top = 99999
+    top_bottom = -99999
+    # search search through the list with units
+    for unit in list_with_units:
+        if unit.is_selected:
+            number_of_selested_units += 1
+            # if unit.body_radius > biggest_unit_radius: biggest_unit_radius = unit.body_radius
+            if unit.v_max < slowest_unit_speed: slowest_unit_speed = unit.v_max
+            if unit.coord[0] < top_left: top_left = unit.coord[0]
+            if unit.coord[0] > top_right: top_right = unit.coord[0]
+            if unit.coord[1] < top_top: top_top = unit.coord[1]
+            if unit.coord[1] > top_bottom: top_bottom = unit.coord[1]
+    
+    coord_of_squad = [top_left + (top_right - top_left) / 2, top_top + (top_bottom - top_top) / 2]
+    translation_vector = [target[0] - coord_of_squad[0], target[1] - coord_of_squad[1]]
+
+    # set new target
+    for unit in list_with_units:
+        if unit.is_selected:
+            if is_ctrl_down:
+                unit.base.movement_target.append([unit.coord[0] + translation_vector[0], unit.coord[1] + translation_vector[1]])
+                unit.set_v_max_squad(slowest_unit_speed)
+            else:
+                unit.base.movement_target = [[unit.coord[0] + translation_vector[0], unit.coord[1] + translation_vector[1]]]
+                unit.set_v_max_squad(slowest_unit_speed)
+
+
+def set_new_target_regroup(list_with_units, target, is_ctrl_down):
+# set new movement target to all selected units
+# units will regroup into spiral
     number_of_selested_units = 0
     current_unit = 0
     biggest_unit_radius = 0